SampleUI project, Tree/TreePanel UI elements

+++ Source/Engine/inc/Animation/Animation.h
@@ -38,7 +38,7 @@ public:
     void Hide();
     void ToggleVisible();
 
-    void SetPlaybackSpeed(double speed);
+    void SetPlaybackSpeed(f32 speed);
 
 private:
     const Animation* m_Animation;
